staticx: Add -l option to specify extra libraries
There may be some libraries that are required by the application, but are
not detected by 'ldd', because e.g. they are loaded with dlopen().

For example, PyInstaller's libpython links against libpthread, which
because it is closely tied to libc, PyInstaller doesn't include in its
archive. So we can include it with "-l /lib64/libpthread.so.0". In the
future, we'll probably add a "hook" for PyInstalled applications where we
actually look inside their archive and see if there's anything else we
need to bring along.