staticx: Process PyInstaller archive and add shared lib dependencies
We use the PyInstaller CArchiveReader API to determine if the given
application is a PyInstaller archive, and if so, dump out all of the shared
libraries to a temporary directory. We then run 'ldd' on those libraries,
and add any libraries not already present in our archive or the PyInstaller
archive.

Closes #13