Implement auto-versioning using Git
If setup.py exists, we assume we're running from source, and attempt to
get the repository version and append it to the hard-coded base version.
If a .git directory exists, we'll use 'git describe' to get the most
detailed information. Otherwise, we'll get the information from a string
which git will populate during 'git archive'.

With no setup.py, we'll assume we're installed, and we'll consult
pkg_resources.

This scheme was adapted and simplified from that of PyInstaller
https://github.com/pyinstaller/pyinstaller/