setup.py: Detect TRAVIS_BUILD_NUMBER
PyPI doesn't like PEP 440 local names. Normally, this isn't a big deal
because we only deploy tags to pypi, so we have no local component in
the version. For test pypi however, we deploy on every push to 'master',
so we need additional version information. Instead of using the git repo
information like we'd prefer, we'll append the TRAVIS_BUILD_NUMBER as
another dotted component of the version.