pyinstaller: Set libpath when determining shobj deps
This emulates the behavior of the PyInstaller bootloader:

https://github.com/pyinstaller/pyinstaller/blob/v3.4/bootloader/src/pyi_utils.c#L808-L811

This is necessary because some shared objects may depend on other shared
objects in the PyInstaller archive. Without this change, Staticx would
still build a valid program (because there are no additional
dependencies in this case), but ldd would report a missing library.

Fixes additional cases reported in #61.