bootloader: Remove readlinka()
This implementation followed the guidance of the readlink(2) man page and
used the st_size result of lstat() to determine the size of the buffer to
allocate. Unfortunately, this doesn't work on the /proc filesystem, where
st_size is zero.

This function isn't used and can easily be replaced with realpath() so
let's simply remove it.