ci: Simplify .travis.yml and consistently use xenial
PR #66 moved almost everything from trusty to xenial, but Python 3.4
needed to stay on trusty. This must have been fixed at some point because
Scuba can use 3.4 on xenial:

https://github.com/JonathonReinhart/scuba/blob/0244c81ec4/.travis.yml

Not only is this simpler, but it also fixes a problem I was having where
"realpath" wasn't available on trusty, but was on xenial.