Change PROG_FILENAME to be a symlink to the originally-named user program
Previously, the user program would always be added to the archive as
".staticx.prog" and executed as such. This would cause problems for
programs that change their behavior based on the value of basename(argv[0])
(e.g. busybox).

Now, we add the program with its original name to the archive, and
PROG_FILENAME is instead a symlink to the program. This symlink is not
executed directly (as that would not solve the problem), but rather used
by the bootloader to identify the name of the user program to execute.

Fixes #134