staticx: Keep LD_LIBRARY_PATH when invoking ldd
This phase of staticx, library path discovery, is intended to run in
whatever environment the user has created. That means we should keep the
value of LD_LIBRARY_PATH when invoking ldd.

The existing ('if libpath') already handled the case where LD_LIBRARY_PATH
is in the env and properly prepended to it. But LD_LIBRARY_PATH could
never be in the empty env.

This change simply keeps LD_LIBRARY_PATH.

Fixes #184