staticx: Don't add symlinks with the same name as target
Background:
The archive contains a flat collection of all dependent libraries.
We also add symlinks as necessary for library versions
(e.g. libfoo.so.1 -> libfoo.so.1.2 -> libfoo.so.1.2.3).

Problem:
The code assumed that a library symlink would only ever point to a
library (or another symlink) with a different basename.  It did not
consider the case where a symlink would point to a library with the
same name (but in a different directory). If we would add such a
symlink in our flat archive directory, it would be self-referential.
The actual crash was a check in the staticx python code, which prevented
ELOOP (Too many levels of symbolic links) at runtime.

Solution:
Skip the creation of any symlinks with the same name as their target, as
those are pointless in our flat directory.