staticx: Don't raise exception from is_dynamic_elf
Previously, if a non-ELF file appeared in a PyInstaller archive as a
"binary" file, staticx would crash when open_elf() raised an
InvalidInputError.

Rename is_dynamic() to is_dynamic_elf() to make it clear that it now
returns False if the file is not an ELF file at all.

Fixes #268.