staticx: Add type info to staticx.version
This updates git_describe() to return Optional[GitDescribe] to simplify
typing.

The whole file looks much closer now to this (and in some ways better):
https://github.com/JonathonReinhart/scuba/blob/18135bb19a/scuba/version.py

This also adds importlib-metadata to requirements.txt for development.
This is because even though we're running with Python 3.11 on Debian 12,
we specify python_version = "3.8" in pyproject.toml, which means the
backport is needed for type checking.