staticx: Add type info to staticx.api
One slightly annoying issue is that I like to set attributes to None in
close() after they've been closed. But this requires marking their types
as Optional, which requires checking that they are not None where they
are used. A well-polished API would check and raise a "closed"
exception. For now, we just assert.

Also remove wildcard imports.

Note that there is a circular dependency between staticx.api and
staticx.hooks. Because of this, we need to add ignore[no-untyped-call]
to the staticx.hooks.run_hooks() call until types are added there (which
cannot happen until after this commit).

This also fixes #278.