staticx: Add type info to staticx.api WIP
One slightly annoying issue is that I like to set attributes to None in
close() after they've been closed. But this requires marking their types
as Optional, which requires checking that they are not None where they
are used. A well-polished API would check and raise a "closed"
exception. For now, we just assert.

This also fixes #278.